don't throw exceptions when trying to delete a file that doesn't exist

+++ src/backendMain/kotlin/edu/duke/bartesaghi/micromon/io.kt
@@ -184,8 +184,11 @@ fun Path.createDirsIfNeeded() = apply {
 	}
 }
 
-fun Path.delete() =
-	Files.delete(this)
+fun Path.delete() {
+	if (exists()) {
+		Files.delete(this)
+	}
+}
 
 fun Path.deleteDirRecursively() = apply {
 	if (exists() && isDirectory()) {
